# public/swagger-spec.json
{"openapi":"3.0.0","paths":{"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDTO"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/hello":{"get":{"operationId":"HelloController_fetch","parameters":[{"name":"role","required":true,"in":"query","schema":{"enum":["admin","user"],"type":"string"}},{"name":"id","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":"哈哈哈哈"}},"tags":["hello"]},"post":{"operationId":"HelloController_save","parameters":[],"requestBody":{"required":true,"description":"输入name","content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"201":{"description":""}},"tags":["hello"]},"delete":{"operationId":"HelloController_delete","parameters":[],"responses":{"200":{"description":""}},"tags":["hello"]}},"/hello/{id}":{"patch":{"operationId":"HelloController_edit","parameters":[{"name":"id","required":true,"in":"path","schema":{}}],"requestBody":{"required":true,"description":"输入name","content":{"application/json":{"schema":{"type":"string"}}}},"responses":{"200":{"description":""}},"tags":["hello"]}},"/users":{"get":{"operationId":"UsersController_findOne","parameters":[],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]},"post":{"operationId":"UsersController_regist","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterInfoDTO"}}}},"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/users/all":{"get":{"operationId":"UsersController_findAll","parameters":[{"name":"size","required":false,"in":"query","schema":{}},{"name":"page","required":false,"in":"query","schema":{}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/users/many":{"post":{"operationId":"UsersController_createMany","parameters":[],"responses":{"201":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}}},"info":{"title":"nest-starter api document","description":"nest-starter api document description","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"RegisterInfoDTO":{"type":"object","properties":{}}}}}